fix: revise Download Links to only download currently visible URLs fr… (#2089)
* fix: revise Download Links to only download currently visible URLs from the user page

* fix: revise Download Links to only download currently visible URLs from the user page
